Add AltReal
