+++ R-package/src/init.c
@@ -119,13 +119,18 @@ SEXP savvy_set_name_external__impl(SEXP x, SEXP name) {
     return handle_result(res);
 }
 
-SEXP savvy_init_altrep_class__impl(DllInfo* dll_info) {
-    SEXP res = savvy_init_altrep_class__ffi(dll_info);
+SEXP savvy_altint__impl(void) {
+    SEXP res = savvy_altint__ffi();
     return handle_result(res);
 }
 
-SEXP savvy_altint__impl(void) {
-    SEXP res = savvy_altint__ffi();
+SEXP savvy_altreal__impl(void) {
+    SEXP res = savvy_altreal__ffi();
+    return handle_result(res);
+}
+
+SEXP savvy_init_altrep_class__impl(DllInfo* dll_info) {
+    SEXP res = savvy_init_altrep_class__ffi(dll_info);
     return handle_result(res);
 }
 
@@ -514,6 +519,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"savvy_get_name_external__impl", (DL_FUNC) &savvy_get_name_external__impl, 1},
     {"savvy_set_name_external__impl", (DL_FUNC) &savvy_set_name_external__impl, 2},
     {"savvy_altint__impl", (DL_FUNC) &savvy_altint__impl, 0},
+    {"savvy_altreal__impl", (DL_FUNC) &savvy_altreal__impl, 0},
     {"savvy_get_class_int__impl", (DL_FUNC) &savvy_get_class_int__impl, 1},
     {"savvy_get_names_int__impl", (DL_FUNC) &savvy_get_names_int__impl, 1},
     {"savvy_get_dim_int__impl", (DL_FUNC) &savvy_get_dim_int__impl, 1},
