+++ R-package/src/rust/api.h
@@ -15,8 +15,9 @@ SEXP savvy_list_with_names_and_values__ffi(void);
 SEXP savvy_external_person_new__ffi(void);
 SEXP savvy_get_name_external__ffi(SEXP x);
 SEXP savvy_set_name_external__ffi(SEXP x, SEXP name);
-SEXP savvy_init_altrep_class__ffi(DllInfo* dll_info);
 SEXP savvy_altint__ffi(void);
+SEXP savvy_altreal__ffi(void);
+SEXP savvy_init_altrep_class__ffi(DllInfo* dll_info);
 SEXP savvy_get_class_int__ffi(SEXP x);
 SEXP savvy_get_names_int__ffi(SEXP x);
 SEXP savvy_get_dim_int__ffi(SEXP x);
