+++ R-package/src/rust/src/altrep.rs
@@ -1,5 +1,7 @@
-use savvy::altrep::{register_altinteger_class, AltInteger};
+use savvy::altrep::{register_altinteger_class, register_altreal_class, AltInteger, AltReal};
 use savvy::savvy;
+
+// integer
 
 struct MyAltInt(Vec<i32>);
 impl savvy::IntoExtPtrSexp for MyAltInt {}
@@ -24,14 +26,48 @@ impl AltInteger for MyAltInt {
 }
 
 #[savvy]
-fn init_altrep_class(dll_info: *mut savvy::ffi::DllInfo) -> savvy::Result<()> {
-    register_altinteger_class::<MyAltInt>(dll_info)?;
-    Ok(())
+fn altint() -> savvy::Result<savvy::Sexp> {
+    let v = MyAltInt::new(vec![1, 2, 3]);
+    let v_altrep = v.into_altrep()?;
+    Ok(savvy::Sexp(v_altrep))
+}
+
+// real
+
+struct MyAltReal(Vec<f64>);
+impl savvy::IntoExtPtrSexp for MyAltReal {}
+
+impl MyAltReal {
+    fn new(x: Vec<f64>) -> Self {
+        Self(x)
+    }
+}
+
+impl AltReal for MyAltReal {
+    const CLASS_NAME: &'static str = "MyAltReal";
+    const PACKAGE_NAME: &'static str = "TestPackage";
+
+    fn length(&mut self) -> usize {
+        self.0.len()
+    }
+
+    fn elt(&mut self, i: usize) -> f64 {
+        self.0[i]
+    }
 }
 
 #[savvy]
-fn altint() -> savvy::Result<savvy::Sexp> {
-    let v = MyAltInt::new(vec![1, 2, 3]);
-    let v_altrep = savvy::altrep::create_altrep_instance(v)?;
+fn altreal() -> savvy::Result<savvy::Sexp> {
+    let v = MyAltReal::new(vec![1.0, 2.0, 3.0]);
+    let v_altrep = v.into_altrep()?;
     Ok(savvy::Sexp(v_altrep))
+}
+
+// initialization
+
+#[savvy]
+fn init_altrep_class(dll_info: *mut savvy::ffi::DllInfo) -> savvy::Result<()> {
+    register_altinteger_class::<MyAltInt>(dll_info)?;
+    register_altreal_class::<MyAltReal>(dll_info)?;
+    Ok(())
 }
